Add optional timeout argument to open and open_simple
While 'search' accepts a timeout option, 'open' and 'open_simple' do not.
This is a problem if there is a network failure; 'open' can hang for a
long time.

That might be fine for small scripts and such, but for production
services, being able to bound how long your service will wait for a
connection before giving up is essential.